+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Unreleased
 - Moved functionality into `sys` module
 - Introduced `MatrixStack` type
 - Introduced `Shader` and `Program` types
+- Introduced `Texture` type and supporting infrastructure
 
 
 0.1.0